Move providers to their own directory now that models no longer exist
Also did some minor refactoring and strong typing here and there in legacy code.
