+++ lib/providers/base/operation_result.dart
