+++ lib/widgets/atoms/reminder_banner.dart
@@ -3,7 +3,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:mm_flutter_app/constants/app_constants.dart';
 import 'package:provider/provider.dart';
 
-import '../../data/models/user/user_provider.dart';
+import '../../providers/user_provider.dart';
 
 class ReminderBanner extends StatelessWidget {
   final String titleText;
