+++ lib/widgets/screens/channel_messages/message_bubble/message_hoverover.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mm_flutter_app/data/models/user/user_provider.dart';
+import 'package:mm_flutter_app/providers/user_provider.dart';
 import 'package:provider/provider.dart';
 
 import 'message_bubble.dart';
