+++ lib/widgets/screens/channel_messages/reply_message.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:mm_flutter_app/data/models/user/user.dart';
-import 'package:mm_flutter_app/data/models/user/user_provider.dart';
+import 'package:mm_flutter_app/providers/channels_provider.dart';
+import 'package:mm_flutter_app/providers/user_provider.dart';
 import 'package:provider/provider.dart';
 
-import '../../../data/models/messages/messages_provider.dart';
+import '../../../providers/messages_provider.dart';
 
 class ReplyMessage extends StatelessWidget {
   const ReplyMessage(
@@ -12,7 +12,7 @@ class ReplyMessage extends StatelessWidget {
       required this.participants,
       this.onClose})
       : super(key: key);
-  final List participants;
+  final List<ChannelParticipant> participants;
   final ChannelMessage replyMessage;
   final VoidCallback? onClose;
 
@@ -21,9 +21,16 @@ class ReplyMessage extends StatelessWidget {
   }
 
   String _participantName({userId}) {
-    final User participant = participants
-        .firstWhere((item) => item.id == userId, orElse: () => null);
-    return participant.fullName!.trim().split(RegExp(' +')).take(1).join();
+    final ChannelParticipant? participant =
+        participants.where((item) => item.user.id == userId).firstOrNull;
+    if (participant != null) {
+      return participant.user.fullName!
+          .trim()
+          .split(RegExp(' +'))
+          .take(1)
+          .join();
+    }
+    return '';
   }
 
   @override
