+++ lib/widgets/screens/home_menu/home_menu.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mm_flutter_app/data/models/user/user_provider.dart';
+import 'package:mm_flutter_app/providers/user_provider.dart';
 import 'package:mm_flutter_app/utilities/utility.dart';
 import 'package:mm_flutter_app/widgets/screens/user_settings/user_settings.dart';
 import 'package:provider/provider.dart';
