+++ lib/widgets/screens/messages/messages_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:mm_flutter_app/data/models/channels/channels_provider.dart';
+import 'package:mm_flutter_app/providers/channels_provider.dart';
 import 'package:provider/provider.dart';
 
-import '../../../data/models/user/user_provider.dart';
+import '../../../providers/user_provider.dart';
 import 'messages_list.dart';
 
 class ChannelsScreen extends StatelessWidget {
