+++ lib/widgets/screens/profile/your_email.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mm_flutter_app/__generated/schema/schema.graphql.dart';
-import 'package:mm_flutter_app/data/models/user/user_provider.dart';
+import 'package:mm_flutter_app/providers/user_provider.dart';
 import 'package:provider/provider.dart';
 
 import '../../atoms/text_form_field_widget.dart';
