+++ lib/widgets/screens/user_settings/user_settings.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:mm_flutter_app/providers/user_provider.dart';
 import 'package:provider/provider.dart';
-
-import 'package:mm_flutter_app/data/models/user/user_provider.dart';
 
 class UserSettings extends StatefulWidget {
   const UserSettings({super.key});
